Fixed video read queue issues

+++ include/image_ffmpeg.h
@@ -93,7 +93,7 @@ class Image_FFmpeg : public Image
 
     // Frame size history, used to keep the frame buffer smaller than _maximumBufferSize
     std::vector<int64_t> _framesSize{};
-    int64_t _maximumBufferSize{1 << 29};
+    int64_t _maximumBufferSize{(int64_t)1 << 29};
 
     std::mutex _videoQueueMutex;
     std::mutex _videoSeekMutex;
